Fixed custom logging to not match itself
This is quite a funny mistake.
The command reads the boxstarter log file and matches all the line containing the word 'Failures', but that log file also includes the call of the command itself... so the resulting log file for the 'Failures' ends up containing copies of the command that was used to create the failures log.
Hilarious accidental recursion, almost